Page JSF et backing bean transfert d'argent d'un compte à l'autre

+++ src/main/java/mg/itu/randrianarivelo/ejb/GestionnaireCompte.java
@@ -51,7 +51,7 @@ public class GestionnaireCompte {
         return em.merge(compteBancaire);
     }
     
-    public CompteBancaire findById(int id){
+    public CompteBancaire findById(long id){
         return em.find(CompteBancaire.class, id);
     }
 }
